Scroll current file to center.
At the moment, the javascript scrolls the current file in the left
navbar to the top of the screen, which I find a bit irritating because
then you can't see the files above it anymore.  This PR changes the
behavior so that it is scrolled into the vertical center of the screen.