Gray-out implicit arguments instead of collapsing them
At the moment, implicit arguments (including instance-implicit
arguments) are collapsed to `{…}`.  This approach impacts not only
discoverability but it also obscures the theorem statement by omitting
the type-class requirements.

It is particularly annoying that the type-class arguments are not shown
by default, because these are typically critical to understand the
theorem statement.

As an easy fix, this PR proposes to *gray-out* the implicit arguments
instead of programatically hiding them behind an unintuitive `{…}`
button.  As soon as the user hovers over the arguments, the arguments
are shown in full color.  This is done in an eye-pleasing way by setting
appropriate easing animations.