use class instead of dict for structure field names
This way pylint will notice if you accidentally use the wrong key.
I'm not sure if this is officially "pythonic".

It seems like more could be done translating the json to actual data structures with named fields; then these lookup tables wouldn't be necessary!
But this might be a lot of work for little gain.